This commit was generated by cvs2svn to compensate for changes in r18681, which included commits to RCS files with non-trunk default branches.
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@18682 728af825-273c-0410-89f9-f7d3b574a069

+++ third/gnome-media/cddb-slave2/cddb-slave.c
@@ -225,7 +225,7 @@ static void
 do_goodbye (ConnectionData *cd)
 {
 	char *quit;
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 	
 	/* Send quit command */
@@ -383,7 +383,7 @@ do_read (ConnectionData *cd,
 	 const char *dtitle)
 {
 	char *query;
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 	
 	/* Send read command */
@@ -674,7 +674,7 @@ static void
 do_query (ConnectionData *cd)
 {
 	char *query;
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 	
 	/* Send query command */
@@ -744,7 +744,7 @@ static void
 do_hello (ConnectionData *cd)
 {
 	char *hello;
-	gsize bytes_writen;
+	guint bytes_writen;
 	GIOError status;
 	
 	/* Send the Hello command
